+++ Sources/MapboxNavigation/AVAudioSession.swift
@@ -1,0 +1,27 @@
+import AVFoundation
+
+@available(*, deprecated, message: "This extension is no longer supported.")
+public extension AVAudioSession {
+    
+    // MARK: Adjusting the Volume
+    
+    func tryDuckAudio() -> Error? {
+        do {
+            try setCategory(.playback, mode: .voicePrompt, options: [.duckOthers, .mixWithOthers])
+            try setActive(true)
+        } catch {
+            return error
+        }
+        return nil
+    }
+    
+    func tryUnduckAudio() -> Error? {
+        do {
+            try setActive(false,
+                          options: [.notifyOthersOnDeactivation])
+        } catch {
+            return error
+        }
+        return nil
+    }
+}